+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## current
 - Added active flag to players/goalies
+- Clear rookie flag on new (regular) season
 
 ## 0.10.0
 - Don't show player number zero in reports